External Libraries: Backport upstream PHP 8.5 fixes for getID3.
In absence of a new release, this cherry-picks [https://github.com/JamesHeinrich/getID3/commit/8cb292333f60be1c0a2ba65acbfdb5f06f174536 8cb29233] and [https://github.com/JamesHeinrich/getID3/commit/dbda40de6dae134a1381452a1e78ea9d02d895c2 dbda40de].

Props swissspidy, vidugupta, TobiasBg, desrosj.
Fixes #64051. See #63061.
Built from https://develop.svn.wordpress.org/trunk@60812


git-svn-id: http://core.svn.wordpress.org/trunk@60148 1a063a9b-81f0-0310-95a4-ce76da25c4cd

+++ wp-includes/ID3/module.audio.ogg.php
@@ -787,29 +787,29 @@ $this->warning('Ogg Theora (v3) not fully supported in this version of getID3 ['
 				switch ($index) {
 					case 'rg_audiophile':
 					case 'replaygain_album_gain':
-						$info['replay_gain']['album']['adjustment'] = (double) $commentvalue[0];
+						$info['replay_gain']['album']['adjustment'] = (float) $commentvalue[0];
 						unset($info['ogg']['comments'][$index]);
 						break;
 
 					case 'rg_radio':
 					case 'replaygain_track_gain':
-						$info['replay_gain']['track']['adjustment'] = (double) $commentvalue[0];
+						$info['replay_gain']['track']['adjustment'] = (float) $commentvalue[0];
 						unset($info['ogg']['comments'][$index]);
 						break;
 
 					case 'replaygain_album_peak':
-						$info['replay_gain']['album']['peak'] = (double) $commentvalue[0];
+						$info['replay_gain']['album']['peak'] = (float) $commentvalue[0];
 						unset($info['ogg']['comments'][$index]);
 						break;
 
 					case 'rg_peak':
 					case 'replaygain_track_peak':
-						$info['replay_gain']['track']['peak'] = (double) $commentvalue[0];
+						$info['replay_gain']['track']['peak'] = (float) $commentvalue[0];
 						unset($info['ogg']['comments'][$index]);
 						break;
 
 					case 'replaygain_reference_loudness':
-						$info['replay_gain']['reference_volume'] = (double) $commentvalue[0];
+						$info['replay_gain']['reference_volume'] = (float) $commentvalue[0];
 						unset($info['ogg']['comments'][$index]);
 						break;
